requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
black==22.3.0
certifi==2020.11.8
chardet==3.0.4
click==8.1.3
Django==3.1.3
django-cacheops==6.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
funcy==1.16
greenlet==1.1.2
grpcio==1.38.1
grpcio-tools==1.38.1
hashids==1.3.1
idna==2.10
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.0.1
platformdirs==2.5.2
protobuf==3.17.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
qcloud-python-sts==3.0.5
redis==3.5.3
requests==2.25.0
shortuuid==1.0.8
six==1.15.0
SQLAlchemy==1.4.39
sqlparse==0.4.1
tencentcloud-sdk-python==3.0.413
tomli==2.0.1
typing-extensions==4.2.0
tzlocal==2.1
urllib3==1.26.2
uWSGI==2.0.20
